Implemented protection against incorrect settings

+++ TagsCloudVisualization/AppSettings/FontSettings.cs
@@ -5,7 +5,7 @@ namespace TagsCloudVisualization.AppSettings
     public class FontSettings
     {
         public Font Font { get; set; } = new Font("Times New Roman", 14);
-        public float MinSize { get; set; } = 7;
-        public float MaxSize { get; set; } = 55;
+        public uint MinSize { get; set; } = 7;
+        public uint MaxSize { get; set; } = 55;
     }
 }
